VARIORUM-5 #resolve VARIORUM-5 #comment Building the docs has two components: Sphinx and Breathe (and Doxygen, but is not the issue). Sphinx is automatically detected on the system, while Breathe is built from source (updated path to Breathe module). Depending on the version of Sphinx, there may be an incompatibility issue. This host config file points to a more recent version of Sphinx.
Building the docs has two components: Sphinx and Breathe (and Doxygen, but is
not the issue). Sphinx is automatically detected on the system, while Breathe is
built from source (updated path to Breathe module). Depending on the version of
Sphinx, there may be an incompatibility issue. This host config file points to a
more recent version of Sphinx.

Signed-off-by: Stephanie Labasan <labasan1@llnl.gov>